+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**v1.1.0**
+
+- Add `camera.rotate()` for rotations along the z axis.
+- Add `camera.reset()` to reset the view to its original target, distance, and rotation.
+
 **v1.0.0**
 
 - Initial stable version